Sound Component: refactor libcanberra/gsound checks
Currently this is only used in one case; incoming "Find My Phone"
requests, but it's always checked at startup and only once. This caches
the found (or not) backend to be used, while checking if/and when the
action is actually invoked.

This still isn't ideal, because it means we might not find out we can't
use an audible alert until the request happens, but still better and it
should be rare that at least canberra-gtk-play is not availale in GNOME.

Additionally, remove unused play_theme_sound()

fixes #276