Bluetooth: use `bus-type` instead of `connection` for `ChannelService()`
`Gio.DBus.system` (and `session`) is a wrapper for `g_bus_get_sync()`,
while passing a `bus-type` instead will result in `g_bus_get()` (async)
being called for `GAsyncInitable`s.

This means if we fail to get a connection, it will happen in a more
predicatable place *and* we aren't blocking when we don't expect it. We
can probably use this same pattern elsewhere (later).