Mousepad Plugin: abort if sent a Unicode character without Caribou
An upstream bug in AT-SPI (or possibly GJS) cause Xorg to crash,
ultimately ending the user session if `atspi_generate_keyboard_event()`
is passed a Unicode character or keyval.

If a Unicode character is detected and Caribou is not installed, send
a DependencyError notification and abort without sending the remote
device the echo (confirmation) packet.

fixes #323