metadata.json: properly set "version" as a Number-type
We have been setting the "version" property as a String, when it should
be a Number.

Since SweetTooth checks and automatically increments the version by
comparing the last submission in *it's* repository, the fact that it was
being changed from a String to a Number was transparent to us. This has
resulted in the bug fixed by PR #317 twice, but only affected users of
extensions.gnome.org.

*Possibly* related to #316 (?)