Extension: try to avoid problems caused by gnome-shell's #&#xfeff;177
There are a number of problems caused by gnome-shell's quadratic loading
of extensions, especially on Wayland where it can cause a 25s delay
during login, until a call to GetObjects() on our DBusObjectManager
times out.

This change delays activation for 5s after the extension is loaded,
unless the service is already active, in which case there is no change.

fixes #307