FindMyPhone Plugin: refactor to entirely avoid dependency errors
We remove and inline the sound backend from components/sound.js, and
fallback to `gtk_widget_error_bell()` if all else fails.

This means there will never be a dependency error thrown, but depending
on the user settings and system volume, there may never be an audible
sound and the user may not know why. This seems less than ideal, but
it's what users asked for.