Messaging UI: add (initial) placeholder for conversation list
Most devices will sync conversations very quickly and will only see this
for a brief period, however devices that don't support SMS history will
see the 'action-unavailable' icon with a link pointing pointing to #320

In future this should have two lines of text, eg. "No conversations" as
well as the "Help" link. We're too close to release to add new
translatable strings and a proper section should be added to the wiki
help page also.