GtkListBox leak mitigation
There seems to be a mysterious leak happening when all the rows of a
GtkListBox are destroyed. After some brief tests, this doesn't seem to
be related to toggle refs or collection.

To mitigate the worst of it we call `g_object_run_dispose()` instead of
`gtk_widget_destroy()` and invoke the collector anyways. There are some
instances where we can just avoid clearing a list and we should do that
anyways (at some point).