SFTP Plugin: add a method for removing stale host keys
If the remote device changes host keys (eg. Android re-install) we need
to be able to remove stale host keys or the device will always fail to
mount with "Host Key Verification Failure".

We add a method that's called on any unknown mount error since the error
is localized. We only remove the host key for the current IP and port,
which is always in the port range specific to KDE Connect (1739-1764).
Since we always know the host anyways and this is ignored if `sshfs` is
used, this should be exactly as secure (if not more).