Core: consider a NULL byte to be a closed connection
g_data_input_stream_read_line() will return NULL when there is no data
to read, considering it the "end of stream". In practice this seems to
happen occasionally with KDE Connect sockets even when the connection is
not closed, however ignoring the NULL byte can result in the receive
loop spinning out.

We now close the connection when we read a NULL byte, which *may* solve
the mysterious high CPU usage and large memory leak, but comes with the
risk that clients won't be able to connect at all depending on why this
happens for KDE Connect packet streams.

addresses #394