SMS Plugin, Messaging UI: remove auto-notification closing
GSConnect tries to match SMS notifications to conversations and
automatically close them when the conversation is focused.

This hasn't been working well for a very long time, requires an extra
contact lookup and probably won't work at all with group messages. We
can investigate if there's a better way to achieve the same result or if
the functionality can be fixed at a later date.