Nemo compatibility for nautilus-gsconnect.py (#580)
* Add Nemo compatibility to nautilus-gsconnect.py

The extension code will autodetect when it's running in `nemo-python`
(based on the contents of `sys.argv[0]` matching "nemo" and adjust
its imports accordingly. Otherwise, it assumes it's running under
`nautilus-python`.

* Also install file manager extension for Nemo

* Add Nemo extension installation to meson

I was hoping to add 'nemo' as an option alias for 'nautilus', too. But
if there's a way to do that in meson, I couldn't find it in the docs.

* meson: Separate 'nautilus' and 'nemo' options