Lan: fix certificate generation wrt device ids
Sigh. There were two rather inconvenient errors made during v25-26.

A false assumption was made about the hostname being a safe choice for
a unique device ID, which causes problems because all clients appear to
be associating the deviceId with the certificate common name.

Currently, in the case a user runs GSConnect & KDE Connect at different
times on the same system, the hostname would be associated with one
certificate and the second would receive TLS errors and be unable to
pair. This will become far more common when the mDNS backend
materializes and two users on the host can run GSConnect (possibly even
at the same time).

Secondly, when the common name is extracted from the TLS certificate, it
appears a trailing newline (\n) was included in the extracted bytes
which wound up being included in the deviceID.