Clipboard: support clearing the remote clipboard (sort of)
Refactor the clipboard component, XWayland clipboard subprocess and
clipboard plugin to ostensibly allow clearing the remote clipboard.

This will only happen in the case that the current clip/selection owner
disappears or the content is set to a literal empty string (''). In the
case that the clipboard content changes to anything that isn't purely
UTF-8 content, nothing will be synced.

In the case that the clip/selection owner disappears, unlike most
applications that will buffer the content until it's explicitly cleared,
we can't really tell the difference so the remote clipboard will be
cleared. This may result in the remote clipboard being cleared when
globally the desktop still holds pastable content.

In the case that we are taking action to clear the remote clipboard,
all we can actually do is send an empty string literal, which doesn't
technically clear the clipboard on Android, but does effectively erase
any content.

closes #619