Clipboard Component: don't clear clipboard for vanishing selection owner
In Wayland it seems that Mutter uses a temporary selection owner to pass
clipboard content to XWayland. This results in receiving targets/content
very similar to an empty clipboard, but clearing the clipboard in this
case breaks it for all XWayland applications.

Instead we will ignore owner changes that seem to be the selection owner
disappearing, which might result in not clearing the remote clipboard in
some cases, but the alternative is not at all.

cc #619
fixes #677