Extension: call RemoteService.start() if service is enabled
We were calling RemoteService.reload() which triggered the `active`
property of RemoteService, but the "mutex" was still locked because
RemoteService.start() was called in its stack.

Instead, when the extension's enable() hook is invoked, call start() if
the service is enabled and only call reload() if it's disabled (because
it might still be running anyways).

closes #672