Shell Notifications: don't create full proxies for device actions
I don't why we (*cough* I *cough*) did this in the first place, but we
now just make threaded calls to DBus, avoid creating full proxies and
don't auto-start the service.

If the service isn't running, it's probably for a good reason, and this
way we'll get our notifications back if we fail.