SMS Plugin: remove bogus 'insert-address-token' entries from messages
As reported, some MMS messages seem to use a placeholder address which
is later replaced with a proper address. For single-recipient threads
this was confusing the SMS plugin into thinking they were group MMS
which we only support read-only.

Whenever we receive messages, we check and remove any bogus addresses
before proceeding. While we're doing that, we'll include the thread_id
coercion and use a slightly faster for-loop instead of map().

closes #710