Input Component: reimplement prempting the mousepad plugin on Ubuntu
Ubuntu is not shipping Mutter with RemoteDesktop enabled so we need to
disable the mousepad plugin on Wayland, as well as any other system not
supporting the interface.

Ubuntu users will probably have to wait until at least 20.04 or whenever
pipewire makes it into the main repository and they start building
Mutter with the required interface enabled.

closes #688
closes #698