Lan: wipe certificate and unpair on authentication failure
The certificate can be invalid for one of two reasons, but in both cases
we handle it by resetting the certificate and unpairing the device.

If the certificate failed to load, it is probably corrupted or otherwise
invalid. In this case, if we try to continue we will certainly crash the
Android app.

If the certificate did not match what we expected the obvious thing to
do is to notify the user, however experience tells us this is a result
of the user doing something masochistic like nuking the Android app data
or copying settings between machines.