SMS Plugin, Messaging UI: fix undefined/missing addresses
As with addresses with the value `insert-address-token`, undefined or
missing `address` fields in the `addresses` dictionary cause major
problems if passed to the messaging window.

Handle this by replacing the undefined address with the string 'unknown'
so it is treated as a contact not in the address book. This means any
such messages will at least appear to be from the same contact and if
the message is fixed later it will be updated at the next sync.

closes #738

Co-authored-by: ofuangka <ofuangka@users.noreply.github.com>