Gnome 3.36 Support (#778)
* Extension: derive ServiceIndicator from GObject based class

* Extension: remove DoNotDisturb

As of gnome-shell 3.36, there is a built-in do not disturb feature so
there is no reason for us to duplicate this feature. Additionally, this
often caused confusion as many users believed this controlled their
phone's dnd state.

* Extension: fix two minor changes in notification sources

The signal ::acknowledged-changed has been replaced with a proper
GObject property.

NotificationSource.notify() has been renamed showNotification()

* Extension: remove dnd styles

* Extension: update Notification classes to derive from GObject

* Extension: update notificationDaemon methods

* correct some arg names

* Extension Menu: replace removed style class with custom CSS

* Extension: refactor notification banner

* Extension: revert type guard for notification callback

* Extension: misc cleanup to shell/notification.js

* Extension: update PanelMenu.SystemIndicator usage

* GMenu: remove extraneous popup-sub-menu classes

* _gsconnect: misc cleanup

* Remote: avoid a spurious ::g-properties-changed emission after disposal

We do a force disposal when dropping device proxies, but GDBusProxy will
synthesize an emission when the name owner disappears.

This seems to happen after we've disposed it, so add a guard clause to
prevent touching a finalized object.

* Extension: stop pollutiing the global searchPath

We've been polluting the global import searchPath by injecting our
imports into it. Up until now we've just been lucky, but we can just use
the provided Extension object for this.

* metadata.json: bump gnome-shell version to 3.36