Mousepad Plugin: ignore remote keyboard state
Due to some ambiguity in the protocol, the Android app may report the
keyboard as inactive when the plugin preferences still allow for
accepting keyboard events.

Until this can be improved, we'll consider the keyboard state a warning
or notice and still allow sending key presses. In that case it will be up
to the Android app to decide whether to process the events or ignore
them.

cc #892