Core: improve Packet class
Most importantly, avoid deep-cloning when updating or creating
packets from objects.

This is not especially fast or free and reference entanglement
turns out to be not much of an issue. In the rare case a packet
field *might* be assigned by reference, it refers to internal
state that should be updated if it changes before the packet is
sent.