Enable word-wrap in the Remote Keyboard TextView widget
Currently this widget does not have wrap enabled, which means that with very
long text entry, the widget grows in width indefinitely until the user
pressed enter (which also inserts a new line on the remote device). This can
be annoying, as the window may grow to a length which is wider than
the screen.

Instead, this allows long lines of text to wrap indepently of newlines and
makes entering long paragraphs of text much simpler and more pleasant.