Core: repurpose ChannelService:name for something useful
Use the 'name' property to hold the 'deviceName' field of identity
packets and make it writable so it can be updated.

Previously this was used as an arbitrary type string, but this is
largely useless for external consumers which should instead be using the
channel address or type-checking a channel instance when necessary.