Lan Backend: amend the error notification
Amend the error notification for a blocked port from "Network Error" to
"Port already in use", which seems to be the most searched string on
Google for these types of errors.

Since we are looking for a specific error now, check the error code and
pass all other errors to the default error handler.

closes #936