Notification Component: de-duplicate libnotify calls
Since gnome-shell-3.36 the libnotify server is a separate process that
forwards calls to gnome-shell. Our monitoring connection was picking up
both the original call and the forwarded call, resulting in duplicated
notifications.

Check the destination address to ensure we get the proper invocation and
watch the libnotify name owner in case GNOME Shell is restarted.

Co-authored-by: James Westman <flyingpimonster@flyingpimonster.net>