MPRIS Component: migrate players from an interface to a base class
There is an errant emission of `g-properties-changed` that seems to only
occur with `vfunc_g_properties_changed()`. Migrate MPRIS.PlayerProxy to
a composite pattern and use the signal instead.

This also allows migrating MPRIS.Player to a base class instead of an
interface, saving some lines of code when inheriting.