Explicitly place our menu below the network menu
On systems where an extension has placed itself at a higher position
than the network menu (for example using an extension that adds more
controls to the volume slider) the AggregateMenu index of '4' would not
be where you expect it.

This change makes it so that we figure out exactly where the network
menu index is and place ourselves right below it.