Added a getter for Volume.
This should allow the mpris control to start with
current device/playback volume.

Using 0.3 as default instead of 100% so that
because some extensions write to the volume
property in order to get the volume (workaround)
which might cause the volume to suddenly spike to
maximum if the property has not yet been set.
Setting it to 0.3 or 30% ensures that will not
happen while still having enough volume to let
the user know that playback is still ongoing.