WIP: Allow controlling volume for remote players using MPRIS (#1068)
* Fixed Typo: Volume didnt exist causing null value

Changed 'Volume' which didnt exist to '_Volume'. This is the volume being
set / sent out by mpris which we want to convert and send to intended
device. This should fix Volume changes not being sent and causing volume
to spike to 100 (at least on KDEConnect Android app).

Next commit fixes reading the volume from the device. Combination of both
commits will hopefully allow us to remote control media volume using
KDEConnect / GSConnect.

* Added a getter for Volume.

This should allow the mpris control to start with
current device/playback volume.

Using 0.3 as default instead of 100% so that
because some extensions write to the volume
property in order to get the volume (workaround)
which might cause the volume to suddenly spike to
maximum if the property has not yet been set.
Setting it to 0.3 or 30% ensures that will not
happen while still having enough volume to let
the user know that playback is still ongoing.