Lan: make transfers more lenient
If a we receive less than the expected amount of data from a transfer,
we most certainly have corruption. However kdeconnect-android sometimes
unreports a file's size in the payload info.

In that case we will only log a warning and let the user decide if the
file is corrupted when they try to open it. Even though there is a clear
opportunity for abuse, this should not be a security risk because the
device has already passed pairing.

closes #1157