Remove all files when disabling extension
If the extension is uninstalled, the DBus file and others
remain in the system. This can cause problems to other programs
that interact with GSConnect.

This patch fixes this issue by removing all the files when the
extension is disabled, and recreating them when it is enabled.

Fix https://github.com/GSConnect/gnome-shell-extension-gsconnect/issues/1269