Update nautilus-gsconnect.py
Just made it shorter and more readable.  but found out that this script hasn't got any access to the notification system to provide a progress bar.  how would one modify `send_files` to allow how complete is the process.  I have an habit to transmit >100MB files between my computer and phone which takes a few minutes it would be nice to check my notification panel and see how far it is to finishing and maybe an ETA that is updated realtime