crowdin.yml: Change to locale_with_underscore
We've been mapping the translated files to '/po/%locale%.po',
with copious overrides to map many locales to two-letter codes.
That's fine, but whenever Crowdin adds a _new_ language (because
we enabled it), it creates it as e.g. '/po/fi-FI.po'. Basically,
because we told it to.

By using '/po/%locale_with_underscore%.po', we can at least
get it to name the files it creates as e.g. '/po/fi_FI.po',
even without needing an override. And we still have the option
of overriding to a 2-letter code or anything else, just like
we've been doing all along.