Input Component: allow remote input in lockscreen
If the user enables the preference to "Keep Alive" while the screen is
locked, it will be possible to remote input in the lockscreen. This was
implemented by overriding the function that inhibits such access in the
lockscreen (if said preference is enabled)

This was based on the work of the author of the extension "Allow Locked
Remote Desktop", who implemented a very similar feature in his own
program.