Add startup permissions checks
Some files in the extension dir need to be executable for proper
operation, and some packaging methods (GitHub Actions Artifacts)
may create extension bundles with the executable bits removed.
So, on startup we check those files, and make them executable if
necessary. (Only if we're installed as a user extension.)