po: Regenerate LINGUAS file
The contents are now simply the results of running this command:
```sh
ls *.po | sed -e 's/\.po$//' > LINGUAS
```
That should cover every language for which we have translations, and
makes me wonder why we even need that file — can't the build do
effectively the same thing, on-the-fly, when compiling translations?
We should probably look for some way to make that work.