Crowdin: New configuration, default to xx.po names
This is the Crowdin configuration that I _suspected_ was correct
initially, but wimped out on actually making because I was afraid
the changes would be too drastic.

On the strength of Crowdin's own [GetText Project example][1] from
their configuration docs, I'm now even more convinced it's the
_correct_ configuration, and should only require us to create
explicit mappings for the rare translations that **don't** easily
slot into the "use `%two_letters_code%.po` as the filename" default
mapping.

[1]: https://developer.crowdin.com/configuration-file/#example-configurations