nautilus-xtn: Flatten single-device menu
Because Nautilus 43+ uses Gtk4 popovers instead of "real" menus
for its context menu, submenus have to be clicked to be accessed,
they no longer open on hover. This makes a single-item submenu
quite irritating to deal with.

So, when there's exactly one device available as a "Send to..."
target, replace the "Send to Mobile Device > <device_name>" menu
hierarchy with a single "Send to <device_name>" entry.