SFTP plugin: Re-work mount handling
Remove `_listDirectories()` from the SFTP plugin code, and never
try to probe the root of the remote device mount (since it's
always a permission-denied path on Android). Instead, use the
`multiPaths` / `pathNames` keys in the SFTP packet body to
determine the director(y/ies) on the remote device.

Instead of creating a single symlink to the device's root path
in the `by-name` directory, turn the `by-name/${device_name}`
path into a directory, populated on mount with a symlink to each
of the device paths supplied in the packet.

Build the `Files` menu for the device with a list of these shared
path names, each of which (although KDE Connect seems to be back
to just a single shared directory, again) will open the remote path
associated with that share-name. Unmounts (either explicit via the
Files submenu, or observed) will remove these symlinks from the
device directory (which is kept around for future mounts).