MPRIS: Use delay timer to fix update spam
Because the MPRIS plugin listens for metadata changes from players,
and because it will create a new update packet for each change and
immediately send it to the paired device, something like a media
change (advancing in a playlist, for example) can cause 10-15 property
changes in a span of milliseconds, potentially firing a barrage of
10-15 `kdeconnect.mpris` packets at the device. Worse, because each
of these packets contains a _full_ metadata dump, they're 10-15
nearly identical packets, except for a single changed value.

To mitigate this spam, start a 250ms timer running whenever the
`_onPlayerChanged()` listener fires, and only create the update
packet when the timer expires. Ignore any other `_onPlayerChanged`
signals while the timer is running.

Responses to user actions or any `kdeconnect.mpris.request`
packets received will be sent immediately, aborting any running
timer, so responsiveness is not impacted.

Fixes #1919