MPRIS: Increase delay for queueing updates
In some players (*cough* VLC *cough*), on slow systems, the
transition to paused state after a Pause request causes
its canPlay property to return `false` briefly. If our timer
fires too quickly, it'll catch the property in that wrong state
and disable unpausing the player again.

Doubling the timeout so that we don't accidentally catch the
property in the wrong state seems to solve the issue, though
if it comes up again we may need to schedule periodic refreshes
of the properties just to ensure sanity in the exported state
of the players.