Sanitize invalid names instead of rejecting
Inspired by KDE Connect Android, which seems to do the same thing.
If an identity request comes in with an invalid device name, instead
of simply rejecting the connection outright, attempt to remove the
invalid characters from the name. Unless the name is _all_ invalid
characters, allow the device to connect and be identified by the
sanitized version of its name.

Related to #1985