Add Preferences banner for missing openssl
If the `openssl` tool is missing, currently all we do is log a message,
which users are unlikely to see.

Instead, whenever `Gio.TlsCertificate.new_from_paths()` throws an
error (which indicates that `openssl` was not found), set a new
preferences boolean `missing-openssl` to `true`. When that setting
is `true`, the Mobile Settings main window will show a banner message
with information about the issue, as well as a link to the wiki page
about the error.