Address more ESLint problems
Apparently ESLint/JSDoc can no longer see the awaited imports, so
any type names from those used in doc comments have to be quoted
as strings.

Remove Gvc import from systemvolume -- We only needed it for JSDoc.
By the same token, Now that it's no longer used in the doc comments,
ESLint registers Gvc as an unused variable in systemvolume.js. Added
an assert that type-checks the `_mixer` property, just to silence it.
Which is a complete sledgehammer of a fix, but meh.